requirements.txt
pysnow
atlassian-python-api
python-dotenv
schedule